--- src/plugins/data/common/opensearch_query/opensearch_query/custom_filter_matches_index.ts
@@ -1,17 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import { IIndexPattern } from '../../index_patterns';
-import { Filter } from '../filters';
-
-export function customFilterMatchesIndex(filter: Filter, indexPattern?: IIndexPattern | null) {
-  if (!filter.meta?.key || !indexPattern) {
-    return true;
-  }
-  if (filter.meta?.type !== 'custom') {
-    return false;
-  }
-  return filter.meta.index === indexPattern.id;
-}
