+++ src/plugins/data/common/opensearch_query/opensearch_query/filter_matches_index.ts
@@ -31,14 +31,14 @@
 import { IIndexPattern, IFieldType } from '../../index_patterns';
 import { Filter } from '../filters';
 
-/*
- * TODO: We should base this on something better than `filter.meta.key`. We should probably modify
- * this to check if `filter.meta.index` matches `indexPattern.id` instead, but that's a breaking
- * change.
- */
 export function filterMatchesIndex(filter: Filter, indexPattern?: IIndexPattern | null) {
   if (!filter.meta?.key || !indexPattern) {
     return true;
   }
+
+  if (filter.meta?.type === 'custom') {
+    return filter.meta.index === indexPattern.id;
+  }
+
   return indexPattern.fields.some((field: IFieldType) => field.name === filter.meta.key);
 }
